add literalai_util.py
