+++ app/tests/src/evaluation/test_literalai_exporter.py
@@ -1,62 +1,9 @@
 import logging
-from datetime import datetime
 from io import StringIO
-from unittest.mock import MagicMock
 
-import pytest
 from literalai import Step, Thread
-from literalai.my_types import PageInfo
 
-from src.evaluation import literalai_exporter
-from src.evaluation.literalai_exporter import (
-    convert_to_qa_rows,
-    get_project_id,
-    query_threads,
-    save_csv,
-)
-
-THREADS = [Thread(f"th_{i}") for i in range(18)]
-
-
-class MockLiteralAIApi:
-    def __init__(self):
-        self.get_threads_counter = 0
-        self.responses = [THREADS[:5], THREADS[5:10], THREADS[10:]]
-
-    def get_my_project_id(self):
-        return "Test_Project_1234ABC"
-
-    def get_threads(self, *args, **kwargs):
-        threads = self.responses[self.get_threads_counter]
-        self.get_threads_counter += 1
-
-        response = MagicMock()
-        response.data = threads
-        response.totalCount = len(THREADS)
-        response.pageInfo = PageInfo(
-            hasNextPage=threads != self.responses[-1],
-            startCursor=threads[0].id,
-            endCursor=threads[-1].id,
-        )
-        return response
-
-
-@pytest.fixture
-def literalai_client(monkeypatch):
-    mock_lai_client = MagicMock()
-    mock_lai_client.api = MockLiteralAIApi()
-    monkeypatch.setattr(literalai_exporter, "literalai", lambda: mock_lai_client)
-
-
-def test_get_project_id(literalai_client):
-    assert get_project_id() == "Test_Project_1234ABC"
-
-
-def test_query_threads(literalai_client):
-    start_date = datetime.fromisoformat("2025-03-06")
-    end_date = datetime.fromisoformat("2025-03-07")
-    threads = query_threads(start_date, end_date)
-    assert len(threads) == len(THREADS)
+from src.evaluation.literalai_exporter import convert_to_qa_rows, save_csv
 
 
 def create_threads(num: int) -> list[Thread]:
